CHG: refactor the folders by adding hippo folder
