+++ tests/core/layers/test_layers_activation.py
