+++ tests/core/layers/test_layers_linear.py
